# perf(internal/messages): improve performance of broadcasting state