The very first patch release of hashicorp/memberlist introduced a NamedBroadcast type to deal with a significant CPU overhead when queueing many messages for broadcasting in a busy cluster. ckit did not use it. 

Not implementing this interface causes queueing N messages to require N(N + 1) / 2 total calls to Invalidates; for a cluster where 5,000 nodes suddenly join, this is 12.5 million calls.

Implementing this interface makes queueing broadcasts several orders of magnitude faster, as seen in the benchmark results below:

    goos: darwin
    goarch: arm64
    pkg: github.com/grafana/ckit/internal/messages
    cpu: Apple M1 Max
                                 │ queue-before.txt │           queue-after.txt            │
                                 │      sec/op      │    sec/op     vs base                │
    Broadcast/Different_names-10    116215.5n ±  1%   418.8n ±  4%  -99.64% (p=0.000 n=10)
    Broadcast/Same_name-10             137.7n ± 12%   136.0n ± 23%        ~ (p=0.315 n=10)
    geomean                            4.000µ         238.7n        -94.03%

                                 │ queue-before.txt │           queue-after.txt           │
                                 │       B/op       │    B/op      vs base                │
    Broadcast/Different_names-10         82.00 ± 1%   153.50 ± 6%  +87.20% (p=0.000 n=10)
    Broadcast/Same_name-10               72.00 ± 0%    64.00 ± 0%  -11.11% (p=0.000 n=10)
    geomean                              76.84         99.12       +28.99%

                                 │ queue-before.txt │           queue-after.txt            │
                                 │    allocs/op     │ allocs/op   vs base                  │
    Broadcast/Different_names-10         1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
    Broadcast/Same_name-10               2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
    geomean                              1.414        1.000       -29.29%
    ¹ all samples are equal

The CPU performance gains are significant enough that the marginal increase in memory is acceptable; queueing messages for 5,000 nodes will now be ~280x faster while only (briefly) requiring 765KB of memory.